i2v/requirements.txt
dashscope>=1.14.0
flask>=2.0.0
requests>=2.25.0
python-dotenv>=0.19.0
gunicorn>=20.0.0